feat(vm): 实现字节码编译器和虚拟机核心功能
- 添加完整的字节码生成实现，支持各种中间代码指令转换为字节码
- 实现虚拟机调度器(dispatch)，支持字节码指令的解析和执行
- 添加整数哈希函数修复，确保Atom类型的哈希一致性
- 实现常量池管理，支持最多65535个常量的存储和检索
- 添加字节码跳转地址解析和更新机制
- 实现属性对象的getter/setter调用逻辑
- 添加运行时错误检查和边界验证机制
- 完善算术运算、比较操作和控制流指令的执行
- 实现函数调用和返回值处理机制
- 添加性能测试用例验证字节码执行正确性

+++ src/vm/Chunk.cpp
@@ -35,7 +35,7 @@ namespace cial::vm {
 
     [[nodiscard]] String Chunk::dumpConstants(const Runtime *rt) const {
         std::stringstream ss{ "" };
-        size_t i{};
+        u16 i{};
         while(i < _constants.size()) {
             ConstIdx idx{ i };
             ss << fmt::format("{: <6} = {}", idx, dumpConstant(rt, idx));
